Done the contract cleaning up
