+++ components/contracts/contract-card/ContractDetails.tsx
